+++ tests/setup.mjs
@@ -37,7 +37,7 @@ import { popupHtml, mapsLink, spotVisible, sperrWarnung, locApply } from './map.
 import { fbInsights, fbModellCheck, fbRestore, parseFangDatum, beissfensterJetzt, fbRender, fbSortiert, fbCsv, checkFang, fangTeilenText } from './fangbuch.js';
 import { toggleTrip, inTrip, openTrip, tripReady } from './trip.js';
 import { schilfLaden, schilfAus, schilfAktiv } from './reed.js';
-import { computeScore, tagesScore, checkErlaubnisAblauf, openPack } from './tools.js';
+import { computeScore, tagesScore, checkErlaubnisAblauf, openPack, wochenIcs } from './tools.js';
 import { ladeNotiz, speichereNotiz } from './notiz.js';
 import { fullscreenToggle, fullscreenAktiv } from './fullscreen.js';
 import { setzeSicht, aktuelleSicht, ladeSicht, sichtReady } from './sicht.js';
@@ -54,7 +54,7 @@ window.__app = { state, popupHtml, mapsLink, spotVisible, sperrWarnung, locApply
   empfehlung, kandidaten, openPlan, zielfischFor, startzeitFor, tagDatum, tagLabel, bewertZeit, tagesplan, MAX_TAG,
   peilung, himmelsrichtung, winkelDiff, istAuflandig,
   bewerteSpot, bewerteAlle, sterneAus, sterneText, jahreszeit, saison, computeScore, tagesScore,
-  ladeNotiz, speichereNotiz, checkErlaubnisAblauf, openPack,
+  ladeNotiz, speichereNotiz, checkErlaubnisAblauf, openPack, wochenIcs,
   inWindow, inSchonzeit, daysUntilMD, masseAus, fmtMD, fmtDate, solunar, haversine, sunTimes, mondStaerke,
   wtSchaetzung, loadWeather };
 `);